refactor(simint): Moved createTestIdentifiers function to seed.ts file
