+++ packages/alpha/src/__tests__/api/simulatorRoutinesApi.spec.ts
@@ -4,12 +4,12 @@ import { describe, expect, test } from 'vitest';
 import type CogniteClientAlpha from '../../cogniteClient';
 import type { SimulatorRoutineRevision } from '../../types';
 import {
-  createTestIdentifiers,
   getOrCreateDataSet,
   getOrCreateFile,
   setupLoggedInClient,
 } from '../testUtils';
 import {
+  createTestIdentifiers,
   fileExtensionTypes,
   modelTypes,
   routineRevisionConfiguration,
